src/requirements.txt
numpy==1.24.1
Pillow==9.3.0
torch==2.0.0
torchvision==0.15.0
tqdm==4.64.1
python-dotenv==1.0.0
wget==3.2